moved server files to diff dir and trying d3 example
