+++ src/css/styles.scss
@@ -28,6 +28,12 @@
         border: none;
         padding: 10px;
       }
+
+      .chart {
+        height: 300px;
+        width: 800px;
+        border: 1px solid #ccc;
+      }
     }
   }
