+++ src/html/index.html
@@ -2,10 +2,16 @@
 <html>
   <head>
     <title>server test</title>
+
+    <!--  D3  -->
+    <script src="https://d3js.org/d3.v4.min.js"></script>
+
     <!--  in html  -->
     <!--     CSS       -->
+    <link rel="stylesheet" type="text/css" href="css/styles.css">
   </head>
   <body>
+    <div class="chart"></div>
     <ul id="messages"></ul>
     <form action="">
       <input id="m" autocomplete="off" /><button>Send</button>
